Days 6 and 7
Day 6 I only got the first excercise. I ran out of time to work on it
and it was too slow to calculate

Day 7 - part 1 was easy enough using a median. I couldn't think of a
more elegant / math-y way to do the second part, so I just calculated all
the options and got the least expensive one.